// model/model.ts
// benefit props type
export type BenefitType = {
    leftStyle?: boolean;
    title: string;
    text: string;
    navText?: string;
    image1: string;
    image2: string;
}
// feature card props type
export type FeatureCardType = {
    icon?: string;
    title?: string;
    text?: string;
}
// cypto item component props type
export type CryptoItemType = {
    title: string;
    image: string;
    text: string;
}